changed WInnerPrizeRatio to long

+++ src/ALGameGenesisTxMeta.cs
@@ -13,6 +13,6 @@ namespace ADALotto.Models
         public long TicketPrice { get; set; }
         public string TicketPriceUnit { get; set; } = string.Empty;
         public bool IsPermutable { get; set; }
-        public double WinnerPrizeRatio { get; set; }
+        public long WinnerPrizeRatio { get; set; }
     }
 }
